Added memory estimations. Changed parallelization strategy for probability mask. Please note that the new approach will result in a warning during execution on JURECA-DC due to a deprecated OpenMP method (which is required for older OpenMP versions).

+++ src/toolbox.h
@@ -136,9 +136,9 @@ namespace PLImg {
         }
 
         namespace labeling {
-            unsigned long long getLargestAreaConnectedComponentsMemoryEstimation(const cv::Mat& image);
-            unsigned long long getConnectedComponentsMemoryEstimation(const cv::Mat& image);
-            unsigned long long getConnectedComponentsLargestComponentMemoryEstimation(const cv::Mat& image);
+            float getLargestAreaConnectedComponentsMemoryEstimation(const cv::Mat& image);
+            float getConnectedComponentsMemoryEstimation(const cv::Mat& image);
+            float getConnectedComponentsLargestComponentMemoryEstimation(const cv::Mat& image);
 
             /**
              * This method allows to search the largest connected component in an image. This connected component will
